reformat code and cleanup imports
usually I wouldn't do that in forks because of larger conflicts when merging upstream changes, but in this case, the upstream project seems a bit
stalled and no big changes are made to the codebase. On the other hand, I plan to maintain this fork (at least for now xD) and fix everything I can.