feat: Fix bug in game.py timer
The code changes in game.py fix a bug related to the timer functionality. Previously, the timer would continue even if the game instance was no longer active. With this fix, the timer will only send a message and skip the next player if the game instance is still active.